Separate async functions in package
