+++ async/compose.js
@@ -1,0 +1,16 @@
+'use strict'
+var defer = require('./defer'),
+    comp = require('../core/compose')
+
+module.exports =
+function compose(/*args*/){
+  var toArgs = [],
+      fromArgs = arguments,
+      len = fromArgs.length,
+      i = 0
+  for(; i < len; i++){
+    toArgs.push(fromArgs[i])
+    toArgs.push(defer())
+  }
+  return comp.apply(null, toArgs)
+}
