+++ async/defer.js
@@ -1,0 +1,7 @@
+'use strict'
+var delay = require('./delay')
+
+module.exports =
+function defer() {
+  return delay()
+}
