+++ async/delay.js
@@ -1,0 +1,110 @@
+'use strict'
+var Prom = require('any-promise')
+
+module.exports = 
+function delay(wait) {
+  return function(xf){
+    return new Delay(wait, xf)
+  }
+}
+function Delay(wait, xf) {
+  var self = this,
+      task = new DelayTask(wait, xf)
+  self.xf = xf
+  self.task = task
+  self._step = spread(task.step, task)
+  self._result = spread(task.result, task)
+}
+
+Delay.prototype.init = function(){
+  var self = this,
+      task = self.task
+  if(task.resolved){
+    return task.resolved
+  }
+
+  return Prom
+    .resolve(self.xf.init())
+}
+Delay.prototype.step = function(value, input) {
+  var self = this,
+      task = self.task
+  if(task.resolved){
+    return task.resolved
+  }
+
+  return Prom
+    .all([value, input])
+    .then(self._step)
+}
+Delay.prototype.result = function(value){
+  var self = this,
+      task = self.task
+  if(task.resolved){
+    return task.resolved
+  }
+
+  return Prom
+    .all([value])
+    .then(self._result)
+}
+
+function DelayTask(wait, xf){
+  this.wait = wait
+  this.xf = xf
+  this.q = []
+}
+DelayTask.prototype.call = function(){
+  var next = this.q[0]
+  if(next && !next.processing){
+    next.processing = true
+
+    var wait = next.wait
+    if(wait > 0){
+      setTimeout(next.fn, wait)
+    } else {
+      next.fn()
+    }
+  }
+}
+DelayTask.prototype.step = function(value, input){
+  var task = this
+  return new Prom(function(resolve, reject){
+    task.q.push({fn: step, wait: task.wait})
+    task.call()
+
+    function step(){
+      try {
+        resolve(task.xf.step(value, input))
+        task.q.shift()
+        if(task.q.length > 0){
+          task.call()
+        }
+      } catch(e){
+        reject(e)
+      }
+    }
+  })
+}
+DelayTask.prototype.result = function(value){
+  var task = this
+  task.resolved = new Prom(function(resolve, reject){
+    task.q.push({fn: result})
+    task.call()
+    function result(){
+      try {
+        task.q = []
+        resolve(task.xf.result(value))
+      } catch(e){
+        reject(e)
+      }
+    }
+  })
+  return task.resolved
+}
+
+function spread(fn, ctx){
+  return function(arr){
+    return fn.apply(ctx, arr)
+  }
+}
