+++ async/reduce.js
@@ -1,0 +1,2 @@
+'use strict'
+module.exports = require('./_core').reduce
